Tested register and update ibc via cli
